fileutl.h: fix the size of Jnk (write-only var) after 297a12 Support large files
The size of this "write-only" var would not match the type expected by
the applied function on arches where long is just 32 bit.

It appears in an inline method. Let's see whether this means that some
clients of libapt will need to be recompiled (and hence we need to
change the soname)...

Perhaps, we could simply hide this method and see whether it is ever used...

+++ apt/apt-pkg/contrib/fileutl.h
@@ -43,7 +43,7 @@ class FileFd
    
    inline bool Read(void *To,unsigned long long Size,bool AllowEof)
    {
-      unsigned long Jnk;
+      unsigned long long Jnk;
       if (AllowEof)
 	 return Read(To,Size,&Jnk);
       return Read(To,Size);
